49.4.5

+++ server/controllers/groupController.js
@@ -239,6 +239,17 @@ async function uploadGroupAvatar(req, res, next) {
         error: "no file"
       });
     }
+    const mimeType = normalizeMime(req.file.mimetype);
+    assertAllowedAvatar({
+      mimeType,
+      fileName: req.file.originalname,
+      size: req.file.size
+    });
+    assertSafeFileBuffer({
+      buffer: req.file.buffer,
+      mimeType
+    });
+
     await deleteUploadedFile({
       url: group.avatar,
       storageKey: group.avatarStorageKey,
